# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ssomero API",
    "version": "1.0.0",
    "description": "OpenAPI description for Ssomero backend (students, courses, auth, etc.)."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "AssessmentDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "classId": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "maxScore": { "type": "integer" },
          "dueDate": { "type": "string", "format": "date-time" }
        },
        "required": ["id","classId","title","maxScore"]
      },
      "CreateAssessmentDto": {
        "type": "object",
        "properties": {
          "classId": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "maxScore": { "type": "integer" },
          "dueDate": { "type": "string", "format": "date-time" }
        },
        "required": ["classId","title","maxScore"]
      },
      "UpdateAssessmentDto": { "$ref": "#/components/schemas/CreateAssessmentDto" },
      "AnnouncementDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "classId": { "type": "string", "format": "uuid", "nullable": true },
          "title": { "type": "string" },
          "body": { "type": "string" },
          "postedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id","title","postedAt"]
      },
      "CreateAnnouncementDto": {
        "type": "object",
        "properties": {
          "classId": { "type": "string", "format": "uuid", "nullable": true },
          "title": { "type": "string" },
          "body": { "type": "string" }
        },
        "required": ["title"]
      },
      "UpdateAnnouncementDto": { "$ref": "#/components/schemas/CreateAnnouncementDto" },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "integer", "format": "int32" }
        },
        "required": ["id","email","role"]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" },
          "role": { "type": "integer", "format": "int32" }
        },
        "required": ["email","password"]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "role": { "type": "integer", "format": "int32" }
        },
        "required": ["email"]
      },
...
    }
  },
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
...
    "/api/assessments": {
      "get": { "summary":"Get all assessments","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AssessmentDto"}}}}}} },
      "post": { "summary":"Create assessment (Admin or Instructor)","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAssessmentDto"}}}},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssessmentDto"}}}}} }
    },
    "/api/assessments/{id}": {
      "get": { "summary":"Get assessment by id","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssessmentDto"}}}},"404":{"description":"Not found"}} },
      "put": { "summary":"Update assessment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAssessmentDto"}}}},"responses":{"204":{"description":"No Content"},"404":{"description":"Not found"}} },
      "delete": { "summary":"Delete assessment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not found"}} }
    },
    "/api/announcements": {
      "get": { "summary":"Get all announcements","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AnnouncementDto"}}}}}} },
      "post": { "summary":"Create announcement (Admin only)","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAnnouncementDto"}}}},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnnouncementDto"}}}}} }
    },
    "/api/announcements/{id}": {
      "get": { "summary":"Get announcement by id","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnnouncementDto"}}}},"404":{"description":"Not found"}} },
      "put": { "summary":"Update announcement","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAnnouncementDto"}}}},"responses":{"204":{"description":"No Content"},"404":{"description":"Not found"}} },
      "delete": { "summary":"Delete announcement","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not found"}} }
    },
    "/api/users": { "get": { "summary":"Get all users (Admin only)","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}}}}} },
      "post": { "summary":"Create user (registration allowed)","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"400":{"description":"Bad request"}} }
    },
    "/api/users/{id}": { "get": { "summary":"Get user by id","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"404":{"description":"Not found"}} },"put": { "summary":"Update user (Admin only)","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"204":{"description":"No Content"},"404":{"description":"Not found"}} },"delete": { "summary":"Delete user (Admin only)","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not found"}} } },
...
  }
}